fix: issue in 'hide' function which is supposed to close all prompt of the selected form but it does not work. see details above.
Issue
-----

Initial implementation of hide function try to remove all '.formError' child classes of the selected form.
In fact, 'formError' prompts are not part of the form childs but these prompt are located at top of the DOM tree.

Solution
--------

I added a class to make prompts identified to their form parents.
Then I updated the hide function to remove correctly all prompts which corresponds to the right form.