Adding Date Range comparison Given 2 fields that are the dates to be checked, This will ensure first date is before the second date provided the Id's are unique and the names are the same Behaves similar to checkbox in a number of respects.
I don't know if you would want to use this but, I find it useful so here ya go. 
If so I'm also implementing a Date time range comparison and i could pass that along as well. 

And please let me know if my git edict isn't correct as this would be my first commit to something on git and I would not want to be rude or anything.  

Terrance Smith @ The Benefit Company
http://www.thebenefitcompany.com/