added inline prompts
new prompt position "inline" to show prompts inline behind the input
(instead of positioning them absolutely). You can either initialize the
validationEngine with the promptPosition option set to "inline" or by
adding a data-attribute (data-prompt-position="inline") to the input.
By adding an optional "data-prompt-target" attribute you can specify the
id of an element to which the prompt should be appended.