Validating Hidden Fields
In some cases, it is useful to be able to validate hidden fields in a
form.
This is particularly true in the case of jquery-UI tabs, if you have a
single form split over multiple tabs.
By adding an option about whether we want to validate hidden fields in a
form, we can override the default behaviour when we need to, while at
the same time maintaining the original behaviour by default.