Allow custom HTML attribute "data-required-class" to override formError prompt class.
I was doing a project where I needed multi-tiered required fields. So "red" was for primary then make more classes for "secondary / tertiary" etc. So that
when the whole form was validated, the "required" alerts were all different depending on the class defined in the "data-required-class" attribute. Thought
it might be useful to others as well. Also added in support for prettySelect.