Possible fix to issue #608
When validating a form with Ajax field validation and binding the
jqv.field.result to monitor validation changes and alter for example 
the class of the field using jQuery, currently, the status will stay
like there was an error, even if the Ajax check succeeded. This fix
seems to solve the issue without breaking the script logics.

I offer then this possible bugfix.